refactor(conf, plot, transform, types, utils): enhance type annotations, improve import structure, and ensure consistent string formatting

+++ src/elisa/binary_system/transform.py
@@ -108,7 +108,7 @@ class RadialVelocityObserverProperties(SystemProperties):
         if not value > 0:
             msg = f"Invalid value of property `mass_ratio`. Expected > 0, given {value}."
             raise ValueError(msg)
-        return float(FLOAT(value))
+        return float(np.float64(value))
 
     @staticmethod
     def asini(value: Any) -> float:
